Fixes an issue with dependency management in pip
Pip has introduced a new dependency management, which will go live in November. Unfortunately one of our dependencies seems to have a requirement on that new system, so we installed the wrong versions of certain dependencies of tk-toolchain. We are now forcing this new dependency management so tk-toolchain can be installed correctly.